+++ src/components/scenes/WalletDetailsScene.tsx
@@ -23,7 +23,7 @@ import { useSceneScrollHandler } from '../../state/SceneScrollState'
 import { useDispatch, useSelector } from '../../types/reactRedux'
 import { NavigationBase, WalletsTabSceneProps } from '../../types/routerTypes'
 import { coinrankListData, infoServerData } from '../../util/network'
-import { calculateSpamThreshold, convertNativeToDenomination, darkenHexColor, zeroString } from '../../util/utils'
+import { calculateSpamThreshold, convertNativeToDenomination, darkenHexColor } from '../../util/utils'
 import { EdgeCard } from '../cards/EdgeCard'
 import { InfoCardCarousel } from '../cards/InfoCardCarousel'
 import { SwipeChart } from '../charts/SwipeChart'
@@ -90,13 +90,13 @@ function WalletDetailsComponent(props: Props) {
 
   // Fiat Balance Formatting
   const exchangeAmount = convertNativeToDenomination(exchangeDenom.multiplier)(exchangeDenom.multiplier)
-  const fiatRate = mul(exchangeAmount, exchangeRate)
+  const fiatRate = mul(exchangeAmount, exchangeRate ?? 0)
   const fiatRateFormat = `${formatNumber(fiatRate && gt(fiatRate, '0.000001') ? fiatRate : 0, {
     toFixed: gt(fiatRate, '1000') ? 0 : 2
   })} ${fiatCurrencyCode}/${currencyCode}`
 
   const spamThreshold = React.useMemo<string | undefined>(() => {
-    if (spamFilterOn && !zeroString(exchangeRate)) {
+    if (spamFilterOn && exchangeRate === 0) {
       return calculateSpamThreshold(exchangeRate, exchangeDenom)
     }
   }, [exchangeDenom, exchangeRate, spamFilterOn])